requirements.txt
torch
torchvision
pymupdf
img2pdf
pillow
numpy
tqdm
safetensors
transformers==4.45.2
addict
einops
fastapi
uvicorn
python-multipart